requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.2.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
Flask-Session==0.8.0
redis==5.0.8
python-dotenv==1.0.1
